correct output

+++ Loan Calculator/Model.swift
@@ -13,7 +13,7 @@ class ReturnModel{
     var sumCalculated = 0.0
     var months = 0
     var years = 0
-    var monthlyPayment = 0
+    var monthlyPayment = 0.0
 
     
     func monthPercentCalculation(_ yearPercent:Double) -> Double {
@@ -30,12 +30,12 @@ class ReturnModel{
     
 
     
-    func paymentCalculation(amount:Double, monthlyPayment:Int, monthsAmount:Int) -> Double{
+    func paymentCalculation(amount:Double, monthlyPayment:Double, monthsAmount:Int) -> Double{
         return (-Double(amount) + Double(monthlyPayment) * Double (monthsAmount))/(Double(Double(monthlyPayment) * Double (monthsAmount))) * 100
     }
     
-    func moonthlyPaymentCalculation(mortgageAmount:Int,monthPercent:Double,all:Double) -> Int{
-        return Int(Double(mortgageAmount) * monthPercent * all / (all - 1))
+    func moonthlyPaymentCalculation(mortgageAmount:Int,monthPercent:Double,all:Double) -> Double{
+        return (Double(mortgageAmount) * monthPercent * all / (all - 1))
     }
     
     
@@ -44,7 +44,7 @@ class ReturnModel{
         monthPercent = monthPercentCalculation(yearPercent)
         all = allCalculation(monthPercent:monthPercent, monthsAmount:monthsAmount)
         amount = amountCalculation(monthlyPayment:monthlyPayment, monthPercent:monthPercent, all:all)
-        payment = paymentCalculation(amount:amount, monthlyPayment:monthlyPayment, monthsAmount:monthsAmount)
+        payment = paymentCalculation(amount:amount, monthlyPayment:Double(monthlyPayment), monthsAmount:monthsAmount)
 
         return (amount, payment)
     }
@@ -67,13 +67,13 @@ class ReturnModel{
         months = monthsAmount % 12
         years = monthsAmount / 12
         
-        payment = paymentCalculation(amount:amount, monthlyPayment:monthlyPayment, monthsAmount:monthsAmount)
+        payment = paymentCalculation(amount:amount, monthlyPayment:Double(monthlyPayment), monthsAmount:monthsAmount)
         
         return (monthsAmount, months, years, payment)
     }
     
     
-    func fourthReturn(mortgageAmount:Int, monthsAmount:Int, yearPercent:Double) -> (Int, Double){
+    func fourthReturn(mortgageAmount:Int, monthsAmount:Int, yearPercent:Double) -> (Double, Double){
         monthPercent = monthPercentCalculation(yearPercent)
 
         all = allCalculation(monthPercent:monthPercent, monthsAmount:monthsAmount)
